requirements.txt
Django<2
djangorestframework
getrest>=0.2.2
